Move proxy to a separate module
