+++ src/api/mod.rs
@@ -1,1 +1,2 @@
 pub mod files;
+pub mod proxy;
