--- src/client.rs
@@ -1,37 +1,0 @@
-use std::str::FromStr;
-
-use crate::lsp;
-
-// Type to describe a message from the client conveniently.
-pub enum Message {
-    // Valid LSP message
-    Message(lsp::Message),
-    // Invalid JSON
-    Invalid(String),
-    // Close message
-    Close,
-}
-
-// Parse the message and ignore anything we don't care.
-pub async fn filter_map_warp_ws_message(
-    wsm: Result<warp::ws::Message, warp::Error>,
-) -> Option<Result<Message, warp::Error>> {
-    match wsm {
-        Ok(msg) => {
-            if msg.is_close() {
-                Some(Ok(Message::Close))
-            } else if msg.is_text() {
-                let text = msg.to_str().expect("text");
-                match lsp::Message::from_str(text) {
-                    Ok(msg) => Some(Ok(Message::Message(msg))),
-                    Err(_) => Some(Ok(Message::Invalid(text.to_owned()))),
-                }
-            } else {
-                // Ignore any other message types
-                None
-            }
-        }
-
-        Err(err) => Some(Err(err)),
-    }
-}
